ml_project/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
antlr4-python3-runtime==4.8
appnope==0.1.3; platform_system == 'Darwin'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0; python_version >= '3.6'
asttokens==2.0.5
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
beautifulsoup4==4.11.1; python_version >= '3.6'
bleach==5.0.0; python_version >= '3.7'
cffi==1.15.0
debugpy==1.6.0; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints==0.4; python_version >= '3.6'
executing==0.8.3
faker==13.11.1
fastjsonschema==2.15.3
hydra-core==1.1.2
iniconfig==1.1.1
install==1.3.5
ipykernel==6.13.0; python_version >= '3.7'
ipython-genutils==0.2.0
ipython==8.3.0; python_version >= '3.8'
ipywidgets==7.7.0
jedi==0.18.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.1.0; python_version >= '3.6'
jsonschema==4.5.1; python_version >= '3.7'
jupyter-client==7.3.1; python_version >= '3.7'
jupyter-console==6.4.3; python_version >= '3.6'
jupyter-core==4.10.0; python_version >= '3.7'
jupyter==1.0.0
jupyterlab-pygments==0.2.2; python_version >= '3.7'
jupyterlab-widgets==1.1.0; python_version >= '3.6'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.3; python_version >= '3.5'
mistune==0.8.4
nbclient==0.6.3; python_version >= '3.7'
nbconvert==6.5.0; python_version >= '3.7'
nbformat==5.4.0; python_version >= '3.7'
nest-asyncio==1.5.5; python_version >= '3.5'
notebook==6.4.11; python_version >= '3.7'
numpy==1.22.3; python_version >= '3.8'
omegaconf==2.1.2; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pluggy==1.0.0; python_version >= '3.6'
prometheus-client==0.14.1; python_version >= '3.6'
prompt-toolkit==3.0.29; python_full_version >= '3.6.2'
psutil==5.9.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0; os_name != 'nt'
pure-eval==0.2.2
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pygments==2.12.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1; python_version >= '3.7'
pytest==7.1.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
pyyaml==6.0; python_version >= '3.6'
pyzmq==22.3.0; python_version >= '3.6'
qtconsole==5.3.0; python_version >= '3.7'
qtpy==2.1.0; python_version >= '3.7'
scikit-learn==1.1.0
scipy==1.8.0; python_version < '3.11' and python_version >= '3.8'
send2trash==1.8.0
setuptools==62.2.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1; python_version >= '3.6'
stack-data==0.2.0
terminado==0.13.3; python_version >= '3.7'
threadpoolctl==3.1.0; python_version >= '3.6'
tinycss2==1.1.1; python_version >= '3.6'
tomli==2.0.1; python_version >= '3.7'
tornado==6.1; python_version >= '3.5'
traitlets==5.2.0; python_version >= '3.7'
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
